Add files via upload
The first part of this R markdown file uses the tidyverse library to provide an exploratory data analysis on customer credit card data. It uses ggplot2 to create intuitive visualizations on this data.

The second part of this file successfully creates and deploys 3 machine learning algorithms to predict how credit card holders will behave in the future. The data splitting and feature engineering steps are outlined prior to the model creations. The 3 models for this project include: random forest, logistic regression, and KNN. 

An executive summary is included to summarize the project findings at the end of the file.